Corrected answer related to Spring Framework usage in unit tests
