+++ src/test/java/spring/certification/test/q001/example/PropertyResolverTest.java
@@ -1,11 +1,10 @@
-package spring.certification.test;
+package spring.certification.test.q001.example;
 
-import org.junit.Assert;
+import static org.assertj.core.api.Assertions.assertThat;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.springframework.mock.env.MockEnvironment;
-
-import static spring.certification.test.Q001unit.PropertyResolver;
 
 /**
  * Demonstrates usage of Spring Framework support classes used in a unit test.
@@ -13,7 +12,7 @@ import static spring.certification.test.Q001unit.PropertyResolver;
  * @author Valentine Shemyako
  * @since December 31, 2018
  */
-public class Q001unitTest {
+public class PropertyResolverTest {
 
     private static final String SPRING = "Spring";
     private static final String FRAMEWORK = "framework";
@@ -32,7 +31,8 @@ public class Q001unitTest {
     @Test
     public void shouldResolveProperty() {
         PropertyResolver propertyResolver = new PropertyResolver(environment);
-        String value = propertyResolver.resolveProperty(FRAMEWORK);
-        Assert.assertEquals(SPRING, value);
+
+        String actualValue = propertyResolver.resolveProperty(FRAMEWORK);
+        assertThat(SPRING).isEqualTo(actualValue);
     }
 }
